Move some classes from converter-api to xml-converter
