+++ iidm/iidm-xml-converter/src/main/java/com/powsybl/iidm/xml/AbstractNetworkXmlContext.java
@@ -6,7 +6,6 @@
  */
 package com.powsybl.iidm.xml;
 
-import com.powsybl.iidm.AbstractConverterContext;
 import com.powsybl.iidm.anonymizer.Anonymizer;
 
 import java.util.Objects;
