+++ iidm/iidm-xml-converter/src/main/java/com/powsybl/iidm/xml/IidmImportExportMode.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.iidm;
+package com.powsybl.iidm.xml;
 
 /***
  * @author Chamseddine BENHAMED <chamseddine.benhamed at rte-france.com>
