+++ iidm/iidm-xml-converter/src/main/java/com/powsybl/iidm/xml/ImportOptions.java
@@ -4,10 +4,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.iidm.import_;
+package com.powsybl.iidm.xml;
 
 import com.google.common.collect.Sets;
-import com.powsybl.iidm.AbstractOptions;
 
 import java.util.Set;
