+++ iidm/iidm-xml-converter/src/main/java/com/powsybl/iidm/xml/NetworkXmlReaderContext.java
@@ -9,7 +9,6 @@ package com.powsybl.iidm.xml;
 import com.powsybl.commons.extensions.ExtensionXmlSerializer;
 import com.powsybl.commons.xml.XmlReaderContext;
 import com.powsybl.iidm.anonymizer.Anonymizer;
-import com.powsybl.iidm.import_.ImportOptions;
 
 import javax.xml.stream.XMLStreamReader;
 import java.util.*;
