+++ iidm/iidm-xml-converter/src/main/java/com/powsybl/iidm/xml/util/IidmXmlUtil.java
@@ -10,7 +10,7 @@ import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.exceptions.UncheckedXmlStreamException;
 import com.powsybl.commons.extensions.Extension;
 import com.powsybl.commons.xml.XmlUtil;
-import com.powsybl.iidm.export.ExportOptions;
+import com.powsybl.iidm.xml.ExportOptions;
 import com.powsybl.iidm.network.CurrentLimits;
 import com.powsybl.iidm.network.Identifiable;
 import com.powsybl.iidm.network.VoltageLevel;
@@ -106,7 +106,7 @@ public final class IidmXmlUtil {
     /**
      * Assert that the writer context's IIDM-XML version equals or is less recent than a given IIDM-XML version.
      * If not, throw an exception or log an error with a given type of error message,
-     * depending of {@link com.powsybl.iidm.export.ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
+     * depending of {@link ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
      */
     public static void assertMaximumVersion(String rootElementName, String elementName, ErrorMessage type, IidmXmlVersion maxVersion, NetworkXmlWriterContext context) {
         if (context.getVersion().compareTo(maxVersion) > 0) {
@@ -127,7 +127,7 @@ public final class IidmXmlUtil {
     /**
      * Assert that the writer context's IIDM-XML version equals or is less recent than a given IIDM-XML version.
      * If not, throw an exception or log an error with a given type of error message,
-     * depending of {@link com.powsybl.iidm.export.ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
+     * depending of {@link ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
      */
     public static void assertMaximumVersion(String elementName, ErrorMessage type, IidmXmlVersion maxVersion, NetworkXmlWriterContext context) {
         if (context.getVersion().compareTo(maxVersion) > 0) {
@@ -158,7 +158,7 @@ public final class IidmXmlUtil {
     /**
      * Assert that the writer context's IIDM-XML version equals or is more recent than a given IIDM-XML version.
      * If not, throw an exception or log an error with a given type of error message,
-     * depending of {@link com.powsybl.iidm.export.ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
+     * depending of {@link ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
      */
     public static void assertMinimumVersion(String rootElementName, String elementName, ErrorMessage type, IidmXmlVersion minVersion, NetworkXmlWriterContext context) {
         if (context.getVersion().compareTo(minVersion) < 0) {
@@ -179,7 +179,7 @@ public final class IidmXmlUtil {
     /**
      * Assert that the writer context's IIDM-XML version equals or is more recent than a given IIDM-XML version.
      * If not, throw an exception or log an error with a given type of error message,
-     * depending of {@link com.powsybl.iidm.export.ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
+     * depending of {@link ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
      */
     public static void assertMinimumVersion(String elementName, ErrorMessage type, IidmXmlVersion minVersion, NetworkXmlWriterContext context) {
         if (context.getVersion().compareTo(minVersion) < 0) {
@@ -217,7 +217,7 @@ public final class IidmXmlUtil {
      * Assert that the writer context's IIDM-XML version equals or is more recent than a given IIDM-XML version if the value of an attribute or the state of an equipment
      * is not default (interpretable for previous versions).
      * If not, throw an exception or log an error with a given type of error message,
-     * depending of {@link com.powsybl.iidm.export.ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
+     * depending of {@link ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
      */
     public static void assertMinimumVersionIfNotDefault(boolean valueIsNotDefault, String rootElementName,
                                                         String elementName, ErrorMessage type, IidmXmlVersion minVersion,
@@ -246,7 +246,7 @@ public final class IidmXmlUtil {
      * Assert that the writer context's IIDM-XML version equals or is more recent than a given IIDM-XML version if the value of an attribute or the state of an equipment
      * is not default (interpretable for previous versions).
      * If not, throw an exception or log an error with a given type of error message,
-     * depending of {@link com.powsybl.iidm.export.ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
+     * depending of {@link ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
      * If the value is not default and the version is compatible, run a given runnable.
      */
     public static void assertMinimumVersionAndRunIfNotDefault(boolean valueIsNotDefault, String rootElementName,
@@ -271,7 +271,7 @@ public final class IidmXmlUtil {
     /**
      * Assert that the writer context's IIDM-XML version is strictly older than a given IIDM-XML version.
      * If not, throw an exception or log an error with a given type of error message,
-     * depending of {@link com.powsybl.iidm.export.ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
+     * depending of {@link ExportOptions.IidmVersionIncompatibilityBehavior} found in the {@link ExportOptions} of the given context.
      */
     public static void assertStrictMaximumVersion(String rootElementName, String elementName, ErrorMessage type, IidmXmlVersion maxVersion, NetworkXmlWriterContext context) {
         if (context.getVersion().compareTo(maxVersion) >= 0) {
