+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/AbstractXmlConverterTest.java
@@ -7,7 +7,6 @@
 package com.powsybl.iidm.xml;
 
 import com.powsybl.commons.AbstractConverterTest;
-import com.powsybl.iidm.export.ExportOptions;
 import com.powsybl.iidm.network.Network;
 
 import java.io.IOException;
