+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/ExportOptionsTest.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.iidm.export;
+package com.powsybl.iidm.xml;
 
 import com.google.common.collect.Sets;
 import com.powsybl.iidm.network.TopologyLevel;
@@ -14,7 +14,7 @@ import java.nio.charset.StandardCharsets;
 import java.util.HashSet;
 import java.util.Set;
 
-import static com.powsybl.iidm.export.ExportOptions.IidmVersionIncompatibilityBehavior.THROW_EXCEPTION;
+import static com.powsybl.iidm.xml.ExportOptions.IidmVersionIncompatibilityBehavior.THROW_EXCEPTION;
 import static org.junit.Assert.*;
 
 /**
