+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/ImportOptionsTest.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.iidm.import_;
+package com.powsybl.iidm.xml;
 
 import com.google.common.collect.Sets;
 import org.junit.Test;
