+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/LoadingLimitsBugTest.java
@@ -7,7 +7,6 @@
 package com.powsybl.iidm.xml;
 
 import com.powsybl.commons.AbstractConverterTest;
-import com.powsybl.iidm.export.ExportOptions;
 import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
 import org.joda.time.DateTime;
 import org.junit.Test;
